refactor: remove comments
