+++ src/components/cart/CartItem.tsx
